Began building my implementation of a binary search tree
Created a buildTree method that takes an array and returns
a balanced binary tree.
Tested the function by creating a new Tree object.

Created an insert method that takes a value and creates a
new node and adds it to the tree.
Tested the function by calling the function and passing
the value 6.