Added to my implemenation of binary tree
Created a delete method that takes a value and if the value
is present it will delete it.
Tested the function by deleting the root node and see if it
moves the other nodes around to fill in.
Created a find method that finds the value provided of the node
and returns the node.
Tested the function by passing the value of the root node
and see if it returns the correct value of root.